backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
chromadb==0.5.0
httpx==0.27.0
pydantic==2.7.1
tiktoken==0.7.0
tenacity==8.3.0
